+++ canhttp/src/http/tests.rs
@@ -1,8 +1,4 @@
-use crate::http::{
-    HttpConversionLayer, HttpRequestConversionError, HttpRequestConversionLayer, HttpResponse,
-    HttpResponseConversionError, HttpResponseConversionLayer,
-    MaxResponseBytesRequestExtensionBuilder, TransformContextRequestExtensionBuilder,
-};
+use crate::http::HttpConversionLayer;
 use crate::IcError;
 use assert_matches::assert_matches;
 use candid::{Decode, Encode, Principal};
@@ -13,6 +9,8 @@ use ic_cdk::api::management_canister::http_request::{
 };
 use ic_cdk::api::management_canister::http_request::{TransformContext, TransformFunc};
 use tower::{Service, ServiceBuilder, ServiceExt};
+use crate::http::request::{HttpRequestConversionError, HttpRequestConversionLayer, MaxResponseBytesRequestExtensionBuilder, TransformContextRequestExtensionBuilder};
+use crate::http::response::{HttpResponse, HttpResponseConversionError, HttpResponseConversionLayer};
 
 #[tokio::test]
 async fn should_convert_http_request() {
